🐛 fix: Accept branded primitive types as patterns
TS-Pattern v4 used to support branded types, usually defined as an intersection
between a literal type (like `number` or `string`) and an object containing a unique symbol.
There is a bug in v5 that prevents people from using branded types as patterns and this PR fixes it.

This closes issue #167 and #178.